+++ demo/richards/richards_driver.c
@@ -1,5 +1,6 @@
 #include <tdycore.h>
 #include <private/tdycoreimpl.h>
+#include <tdydriver.h>
 #include <tdyts.h>
 #include <tdyio.h>
 
@@ -27,6 +28,7 @@ int main(int argc, char **argv) {
   ierr = PetscOptionsEnd(); CHKERRQ(ierr);
 
   ierr = TDyCreate(&tdy); CHKERRQ(ierr);
+  ierr = TDyDriverInitializeTDy(tdy); CHKERRQ(ierr);
   if (!rank) tdy->io->io_process = PETSC_TRUE;
   tdy->io->print_intermediate = print_intermediate;
   ierr = PrintVec(tdy->U,"initial_solution",-1); CHKERRQ(ierr);
