+++ include/tdyrichards.h
@@ -2,8 +2,10 @@
 #define TDYRICHARDS_H
 
 #include <petsc.h>
+#include <private/tdycoreimpl.h>
 #include <tdycore.h>
 
+PETSC_EXTERN PetscErrorCode TDyRichardsInitialize(TDy);
 PETSC_EXTERN PetscErrorCode TDyRichardsSNESPostCheck(SNESLineSearch,Vec,Vec,Vec,
                                                      PetscBool*,PetscBool*,
                                                      void*);
