+++ src/tdyrichard.c
@@ -1,7 +1,17 @@
-#include <tdypermeability.h>
-#include <tdyporosity.h>
 #include <tdyrichards.h>
-#include <private/tdycoreimpl.h>
+
+PetscErrorCode TDyRichardsInitialize(TDy tdy) {
+  PetscErrorCode ierr;
+  PetscFunctionBegin;
+
+  PetscRandom rand;
+  ierr = PetscRandomCreate(PETSC_COMM_WORLD,&rand); CHKERRQ(ierr);
+  ierr = PetscRandomSetInterval(rand,1.e4,1.e6); CHKERRQ(ierr);
+  ierr = VecSetRandom(tdy->U,rand); CHKERRQ(ierr);
+  ierr = PetscRandomDestroy(&rand); CHKERRQ(ierr);
+
+  PetscFunctionReturn(0);
+}
 
 PetscErrorCode TDyRichardsSNESPostCheck(SNESLineSearch linesearch,
                                         Vec X, Vec Y, Vec W,
